pyproject.toml
[tool.poetry]
name = "ipns-xumm-api"
version = "0.0.1"
description = ""
authors = ["Denis Angell <dangell@transia.co>"]

[tool.poetry.dependencies]
python = "^3.10.6"
connexion = {extras = ["swagger-ui"], version = "^2.13.1"}
Flask-Cors = "^3.0.10"
gunicorn = "^20.1.0"
flask-testing = "^0.8.1"
lmdb = "^1.4.1"
transia-xrpl-py = "^1.7.0a3"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
